# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface CaseSubmission {
  /** @minLength 20 */
  description: string;
  evidences?: string[];
  value?: number | null;
  name: string;
  whatsapp: string;
  email: string;
  state?: string | null;
  city?: string | null;
}

export type CaseResponseStatus =
  (typeof CaseResponseStatus)[keyof typeof CaseResponseStatus];

export const CaseResponseStatus = {
  pending: "pending",
  analyzing: "analyzing",
  active: "active",
  closed: "closed",
} as const;

export interface CaseResponse {
  id: string;
  status: CaseResponseStatus;
  description: string;
  name: string;
  email: string;
  whatsapp: string;
  createdAt: string;
}

export interface LawyerRegistration {
  name: string;
  email: string;
  oab: string;
  phone: string;
  cep?: string | null;
  state?: string | null;
  city?: string | null;
  address?: string | null;
  /** @minLength 8 */
  password: string;
}

export type LawyerResponseStatus =
  (typeof LawyerResponseStatus)[keyof typeof LawyerResponseStatus];

export const LawyerResponseStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface LawyerResponse {
  id: string;
  name: string;
  email: string;
  oab: string;
  status: LawyerResponseStatus;
  createdAt: string;
}

export interface SignInRequest {
  email: string;
  password: string;
}

export interface SignInResponse {
  token: string;
  lawyer: LawyerResponse;
}

export type PaymentRequestMethod =
  (typeof PaymentRequestMethod)[keyof typeof PaymentRequestMethod];

export const PaymentRequestMethod = {
  pix: "pix",
  credit_card: "credit_card",
  debit_card: "debit_card",
} as const;

export interface PaymentRequest {
  caseId: string;
  email: string;
  whatsapp: string;
  cpf: string;
  method: PaymentRequestMethod;
}

export type PaymentResponseStatus =
  (typeof PaymentResponseStatus)[keyof typeof PaymentResponseStatus];

export const PaymentResponseStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
} as const;

export interface PaymentResponse {
  id: string;
  status: PaymentResponseStatus;
  method: string;
  pixCode?: string | null;
  pixQrCode?: string | null;
  amount: number;
  createdAt: string;
}

export interface BlogPost {
  id: string;
  slug: string;
  title: string;
  summary: string;
  content: string;
  category: string;
  imageUrl?: string | null;
  publishedAt: string;
}

export interface BlogPostList {
  posts: BlogPost[];
  total: number;
  page: number;
  limit: number;
}

export type GetBlogPostsParams = {
  page?: number;
  limit?: number;
};