# lib/api-zod/src/generated/types/blogPostList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPost } from "./blogPost";

export interface BlogPostList {
  posts: BlogPost[];
  total: number;
  page: number;
  limit: number;
}