# lib/api-zod/src/generated/types/caseResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseResponseStatus } from "./caseResponseStatus";

export interface CaseResponse {
  id: string;
  status: CaseResponseStatus;
  description: string;
  name: string;
  email: string;
  whatsapp: string;
  createdAt: Date;
}