# lib/api-zod/src/generated/types/caseSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaseSubmission {
  /** @minLength 20 */
  description: string;
  evidences?: string[];
  value?: number | null;
  name: string;
  whatsapp: string;
  email: string;
  state?: string | null;
  city?: string | null;
}