# lib/api-zod/src/generated/types/lawyerRegistration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LawyerRegistration {
  name: string;
  email: string;
  oab: string;
  phone: string;
  cep?: string | null;
  state?: string | null;
  city?: string | null;
  address?: string | null;
  /** @minLength 8 */
  password: string;
}