# lib/api-zod/src/generated/types/lawyerResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LawyerResponseStatus } from "./lawyerResponseStatus";

export interface LawyerResponse {
  id: string;
  name: string;
  email: string;
  oab: string;
  status: LawyerResponseStatus;
  createdAt: Date;
}