# lib/api-zod/src/generated/types/paymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRequestMethod } from "./paymentRequestMethod";

export interface PaymentRequest {
  caseId: string;
  email: string;
  whatsapp: string;
  cpf: string;
  method: PaymentRequestMethod;
}