# lib/api-zod/src/generated/types/paymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentResponseStatus } from "./paymentResponseStatus";

export interface PaymentResponse {
  id: string;
  status: PaymentResponseStatus;
  method: string;
  pixCode?: string | null;
  pixQrCode?: string | null;
  amount: number;
  createdAt: Date;
}