# lib/api-zod/src/generated/types/signInRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procjus Premium API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignInRequest {
  email: string;
  password: string;
}